# services/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "THENAR API",
    "version": "1.0.0",
    "description": "PLAN.md §12 — every route stubbed 501 until its task lands."
  },
  "servers": [
    {
      "url": "/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "invalid_request",
                  "unauthorized",
                  "forbidden",
                  "not_found",
                  "conflict",
                  "unprocessable",
                  "rate_limited",
                  "not_implemented",
                  "internal"
                ]
              },
              "message": {
                "type": "string"
              },
              "details": {}
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/v1/corpora/{id}": {
      "get": {
        "summary": "Get a corpus",
        "x-auth": "public",
        "responses": {
          "501": {
            "description": "not implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/anchors": {
      "get": {
        "summary": "List anchors",
        "x-auth": "public",
        "responses": {
          "501": {
            "description": "not implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/licences/{receiptId}/download": {
      "get": {
        "summary": "Download licensed files",
        "x-auth": "wallet_sig",
        "responses": {
          "501": {
            "description": "not implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}